B-02617: Make CategoryService.storeNewCategory also create unit when creating new top category

+++ modules/cms-core/src/main/java/com/enonic/cms/core/content/category/CategoryXmlCreator.java
@@ -219,7 +219,7 @@ public class CategoryXmlCreator
 
         if ( category.getUnitExcludeDeleted() != null )
         {
-            element.setAttribute( "unitkey", Integer.toString( category.getUnitExcludeDeleted().getKey() ) );
+            element.setAttribute( "unitkey", Integer.toString( category.getUnitExcludeDeleted().getKey().toInt() ) );
         }
 
         element.setAttribute( "subcategories", category.hasChildren() ? "true" : "false" );
