+++ modules/cms-core/src/main/java/com/enonic/vertical/engine/handlers/CategoryHandler.java
@@ -19,7 +19,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import com.enonic.cms.core.CmsDateAndTimeFormats;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
@@ -30,9 +29,7 @@ import com.enonic.esl.util.RelationAggregator;
 import com.enonic.esl.util.RelationNode;
 import com.enonic.esl.util.RelationTree;
 import com.enonic.esl.xml.XMLTool;
-import com.enonic.vertical.engine.AccessRight;
 import com.enonic.vertical.engine.VerticalEngineLogger;
-import com.enonic.vertical.engine.VerticalSecurityException;
 import com.enonic.vertical.engine.XDG;
 import com.enonic.vertical.engine.criteria.CategoryCriteria;
 import com.enonic.vertical.engine.dbmodel.CategoryTable;
@@ -45,18 +42,18 @@ import com.enonic.cms.framework.util.TIntArrayList;
 import com.enonic.cms.framework.xml.XMLDocument;
 
 import com.enonic.cms.core.CalendarUtil;
+import com.enonic.cms.core.CmsDateAndTimeFormats;
 import com.enonic.cms.core.content.ContentKey;
 import com.enonic.cms.core.content.category.CategoryAccessRightsAccumulated;
 import com.enonic.cms.core.content.category.CategoryEntity;
 import com.enonic.cms.core.content.category.CategoryKey;
+import com.enonic.cms.core.content.category.CategoryStatistics;
 import com.enonic.cms.core.content.category.CategoryXmlCreator;
 import com.enonic.cms.core.content.category.access.CategoryAccessResolver;
-import com.enonic.cms.core.security.user.User;
-import com.enonic.cms.store.dao.CategoryDao;
-
-import com.enonic.cms.core.content.category.CategoryStatistics;
 import com.enonic.cms.core.content.contenttype.ContentTypeKey;
+import com.enonic.cms.core.security.user.User;
 import com.enonic.cms.core.security.user.UserEntity;
+import com.enonic.cms.store.dao.CategoryDao;
 
 public class CategoryHandler
     extends BaseHandler
@@ -112,13 +109,13 @@ public class CategoryHandler
         return commonHandler.getInt( sql.toString(), (int[]) null );
     }
 
-    private int[] getCategoryKeysBySuperCategory(CategoryKey superCategoryKey, boolean recursive)
+    private int[] getCategoryKeysBySuperCategory( CategoryKey superCategoryKey, boolean recursive )
     {
 
         return getCategoryKeysBySuperCategories( new int[]{superCategoryKey.toInt()}, recursive );
     }
 
-    public int[] getCategoryKeysBySuperCategories(int[] superCategoryKeys, boolean recursive)
+    public int[] getCategoryKeysBySuperCategories( int[] superCategoryKeys, boolean recursive )
     {
 
         if ( superCategoryKeys == null || superCategoryKeys.length == 0 )
@@ -224,7 +221,8 @@ public class CategoryHandler
             {
                 List<CategoryKey> childrenKeys = category.getChildrenKeys();
                 int[] catKeys = new int[childrenKeys.size()];
-                for ( int j = 0; j < childrenKeys.size(); j++ ) {
+                for ( int j = 0; j < childrenKeys.size(); j++ )
+                {
                     catKeys[j] = childrenKeys.get( j ).toInt();
                 }
                 if ( childrenKeys.size() > 0 )
@@ -313,7 +311,7 @@ public class CategoryHandler
             return -1;
         }
 
-        return category.getUnitExcludeDeleted().getKey();
+        return category.getUnitExcludeDeleted().getKey().toInt();
     }
 
     public boolean hasSubCategories( CategoryKey categoryKey )
@@ -350,7 +348,7 @@ public class CategoryHandler
             }
             if ( adminRead )
             {
-                getSecurityHandler().appendCategorySQL(olduser, sql, true );
+                getSecurityHandler().appendCategorySQL( olduser, sql, true );
             }
 
             con = getConnection();
@@ -367,12 +365,12 @@ public class CategoryHandler
             {
                 subCategories = false;
                 String message = "Failed to count sub-categories. No result given";
-                VerticalEngineLogger.error(message );
+                VerticalEngineLogger.error( message );
             }
 
             if ( !subCategories )
             {
-                int keys[] = getCategoryKeysBySuperCategory(categoryKey, false );
+                int keys[] = getCategoryKeysBySuperCategory( categoryKey, false );
                 for ( int i = 0; !subCategories && i < keys.length; i++ )
                 {
                     subCategories = hasSubCategories( olduser, new CategoryKey( keys[i] ), contentTypeKeys, adminRead );
@@ -382,7 +380,7 @@ public class CategoryHandler
         catch ( SQLException sqle )
         {
             String message = "Failed to count sub-categories: %t";
-            VerticalEngineLogger.error(message, sqle );
+            VerticalEngineLogger.error( message, sqle );
             subCategories = false;
         }
         finally
@@ -463,12 +461,12 @@ public class CategoryHandler
             }
 
             // attribute: created
-            Date createdDate = CmsDateAndTimeFormats.parseFrom_STORE_DATE(root.getAttribute("created"));
+            Date createdDate = CmsDateAndTimeFormats.parseFrom_STORE_DATE( root.getAttribute( "created" ) );
             preparedStmt.setTimestamp( 5, new Timestamp( createdDate.getTime() ) );
 
             // attribute: name
             String name = root.getAttribute( "name" );
-            preparedStmt.setString(6, name);
+            preparedStmt.setString( 6, name );
 
             // element: owner
             Element subElem = (Element) subElems.get( "owner" );
@@ -487,7 +485,7 @@ public class CategoryHandler
                 }
                 else
                 {
-                    preparedStmt.setString(7, description);
+                    preparedStmt.setString( 7, description );
                 }
             }
             else
@@ -509,7 +507,7 @@ public class CategoryHandler
             if ( result == 0 )
             {
                 String message = "Failed to update category.";
-                VerticalEngineLogger.errorUpdate(message, null );
+                VerticalEngineLogger.errorUpdate( message, null );
             }
 
             // set category access rights
@@ -525,12 +523,12 @@ public class CategoryHandler
         catch ( SQLException sqle )
         {
             String message = "Failed to update category: %t";
-            VerticalEngineLogger.errorUpdate(message, sqle );
+            VerticalEngineLogger.errorUpdate( message, sqle );
         }
         catch ( NumberFormatException nfe )
         {
             String message = "Failed to parse a key field: %t";
-            VerticalEngineLogger.errorUpdate(message, nfe );
+            VerticalEngineLogger.errorUpdate( message, nfe );
         }
         finally
         {
@@ -577,15 +575,15 @@ public class CategoryHandler
 
             if ( !getSecurityHandler().validateCategoryUpdate( olduser, categoryDao.findByKey( catKey ).getParent().getKey() ) )
             {
-                VerticalEngineLogger.errorSecurity("User is not allowed to move the category.", null );
+                VerticalEngineLogger.errorSecurity( "User is not allowed to move the category.", null );
             }
 
             if ( !getSecurityHandler().validateCategoryUpdate( olduser, superCatKey ) )
             {
-                VerticalEngineLogger.errorSecurity("User is not allowed to update the new parent category.", null );
+                VerticalEngineLogger.errorSecurity( "User is not allowed to update the new parent category.", null );
             }
 
-            int[] children = getCategoryKeysBySuperCategory(catKey, true );
+            int[] children = getCategoryKeysBySuperCategory( catKey, true );
 
             // Get new unit key
             int unitKey = getUnitKey( superCatKey );
@@ -613,7 +611,7 @@ public class CategoryHandler
         }
         catch ( SQLException e )
         {
-            VerticalEngineLogger.errorUpdate("A database error occurred: %t", e );
+            VerticalEngineLogger.errorUpdate( "A database error occurred: %t", e );
         }
         finally
         {
@@ -687,7 +685,7 @@ public class CategoryHandler
         catch ( SQLException sqle )
         {
             String message = "Failed to count categories: %t";
-            VerticalEngineLogger.error(message, sqle );
+            VerticalEngineLogger.error( message, sqle );
         }
         finally
         {
@@ -948,7 +946,7 @@ public class CategoryHandler
         catch ( SQLException sqle )
         {
             String message = "Failed to get the categories: %t";
-            VerticalEngineLogger.error(message, sqle );
+            VerticalEngineLogger.error( message, sqle );
         }
         finally
         {
@@ -1034,7 +1032,7 @@ public class CategoryHandler
         }
         catch ( SQLException e )
         {
-            VerticalEngineLogger.error("Failed to fetch relation keys: %t", e );
+            VerticalEngineLogger.error( "Failed to fetch relation keys: %t", e );
         }
 
         return doc;
